--- airtable-base-sync/module.yaml
@@ -1,46 +1,0 @@
-moduleId: airtable-table-sync
-
-info:
-  title: Airtable
-  description: Sync an Airtable table to Dataland
-  author: Dataland Devs <support@dataland.io>
-  sourceCodeUrl: https://github.com/dataland-io/dataland-builtin-modules
-  readmePath: README.md
-  iconPath: icon.svg
-
-buildCommand: npm run build
-
-parameters:
-  - airtable-api-key
-  - airtable-base-id
-  - airtable-table-name
-  - airtable-view-name
-  - airtable-allow-writeback-boolean
-  - airtable-dataland-table-name
-  - airtable-fields-list
-
-# Create a .env file in the root of this folder
-# with the following contents:
-# DL_PARAM_AIRTABLE_API_KEY=<your-api-key>
-# DL_PARAM_AIRTABLE_BASE_ID=<your-base-id>
-# DL_PARAM_AIRTABLE_TABLE_NAME=<your-table-name>
-# DL_PARAM_AIRTABLE_VIEW_NAME=<your-view-name>
-# DL_PARAM_AIRTABLE_ALLOW_WRITEBACK_BOOLEAN=<true-or-false>
-# DL_PARAM_AIRTABLE_DATALAND_TABLE_NAME=<your-dataland-table-name>
-# DL_PARAM_AIRTABLE_FIELDS_LIST=<"ALL" or your-fields-list>
-
-workers:
-  - workerId: airtable-table-sync-worker
-    scriptPath: dist/bundle.js
-    triggers:
-      transaction: {}
-      cron:
-        cronExpression: "0 */5 * * * * *"
-    env:
-      AIRTABLE_API_KEY: airtable-api-key
-      AIRTABLE_BASE_ID: airtable-base-id
-      AIRTABLE_TABLE_NAME: airtable-table-name
-      AIRTABLE_VIEW_NAME: airtable-view-name
-      AIRTABLE_ALLOW_WRITEBACK_BOOLEAN: airtable-allow-writeback-boolean
-      AIRTABLE_DATALAND_TABLE_NAME: airtable-dataland-table-name
-      AIRTABLE_FIELDS_LIST: airtable-fields-list
