+++ airtable-multi-table-sync/package.json
@@ -9,6 +9,7 @@
   "dependencies": {
     "@dataland-io/dataland-api-proto": "0.17.0",
     "@dataland-io/dataland-sdk": "0.17.0",
+    "airtable": "^0.11.5",
     "zod": "^3.19.1"
   },
   "devDependencies": {
